feat: rework example pipelines
Common-templates are going to be deprecated in some future version.
This commit is changing example pipelines to remove all automation
related to templates and replaces it with create-vm-from-manifest task.

This change helps to unify pipelines for okd and kubernetes, so now we
can use single version for okd and k8s.

Signed-off-by: Karel Šimon <ksimon@redhat.com>